+++ src/__tests__/AsyncResolver.spec.tsx
@@ -12,12 +12,18 @@ describe('Simple bootstrap tests', () => {
   });
 });
 
-describe('Initial props tests', () => {
+describe('Tests with dummy children', () => {
   it('calls subject exactly once', () => {
     const ObjectResolver = AsyncResolver<object>('subject', {});
 
     const mockResolved = jest.fn();
-    const wrapper = shallow(<ObjectResolver subject={() => new Promise(mockResolved)} />);
+    let wrapper = mount(
+      <ObjectResolver subject={() => new Promise(mockResolved)}>
+        {({ value }) => <span>{value}</span>}
+      </ObjectResolver>,
+    );
+
+    wrapper = wrapper.update();
     expect(mockResolved.mock.calls.length).toBe(1);
 
     wrapper.setProps({ foo: 'bar' });
@@ -28,16 +34,36 @@ describe('Initial props tests', () => {
     const ObjectResolver = AsyncResolver<object>('subject', {});
 
     const mockResolved1 = jest.fn();
-    const wrapper = shallow(<ObjectResolver subject={() => new Promise(mockResolved1)} />);
+    let wrapper = mount(
+      <ObjectResolver subject={() => new Promise(mockResolved1)}>
+        {({ value }) => <span>{value}</span>}
+      </ObjectResolver>,
+    );
+
+    wrapper = wrapper.update();
     expect(mockResolved1.mock.calls.length).toBe(1);
 
     const mockResolved2 = jest.fn();
     wrapper.setProps({ subject: () => new Promise(mockResolved2) });
     expect(mockResolved2.mock.calls.length).toBe(1);
   });
-});
 
-describe('Tests with dummy children', () => {
+  it('pass initial value to children', () => {
+    const ObjectResolver = AsyncResolver<object>('subject', { value: 3 });
+
+    let wrapper = mount(
+      <ObjectResolver subject={() => new Promise(resolve => setTimeout(() => resolve({ value: 4 }), 10))}>
+        {({ value }) => <span>{value}</span>}
+      </ObjectResolver>,
+    );
+
+    wrapper = wrapper.update();
+
+    const span = wrapper.find('span');
+    expect(span.length).toBe(1); 
+    expect(span.text()).toBe('3'); 
+  });
+
   it('can resolve subject and pass result to children', async () => {
     const ObjectResolver = AsyncResolver<object>('subject', { value: 3 });